requirements.txt
requests
flask
bs4
boto3
python-dotenv